requirements.txt
requests==2.32.5
python-dotenv==1.2.1
pandas==3.0.1
folium==0.20.0
matplotlib==3.10.8
